Remove the old ElmBuildAction (superseded by ElmBuildSelectedAction)
